More paranoia re dupes bug
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@12458 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/store/BerkeleyDBFreenetStore.java
@@ -1115,6 +1115,9 @@ public class BerkeleyDBFreenetStore implements FreenetStore {
 		
 		chkDB = environment.openDatabase(null,prefix+"CHK",dbConfig);
 		
+		if(chkDB.count() > 0)
+			throw new IllegalStateException("Wiped old store but it still contains "+chkDB.count()+" keys!");
+		
 		this.fixSecondaryFile = fixSecondaryFile;
 		fixSecondaryFile.delete();
